requirements.txt
memory_profiler
psutil
discord.py==1.0.0a0
praw
peewee
apscheduler
